feat: refactor profile module and service structure; add tests for movie and TV show controllers and services
