+++ src/modules/movie/movie.controller.spec.ts
@@ -1,0 +1,5 @@
+describe("MovieController", () => {
+	it("true is true", () => {
+		expect(true).toBe(true);
+	});
+});
