+++ src/modules/movie/movie.service.spec.ts
@@ -1,0 +1,5 @@
+describe("MovieService", () => {
+	it("true is true", () => {
+		expect(true).toBe(true);
+	});
+});
