+++ src/modules/profile/profile.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
 import { ProfileController } from "./profile.controller";
-import { ProfileService } from "./profile";
+import { ProfileService } from "./profile.service";
 
 @Module({
 	imports: [],
