+++ src/modules/tv-show/tv-show.controller.spec.ts
@@ -1,0 +1,5 @@
+describe("TVShowController", () => {
+	it("true is true", () => {
+		expect(true).toBe(true);
+	});
+});
