+++ src/modules/tv-show/tv-show.service.spec.ts
@@ -1,0 +1,5 @@
+describe("TVShowService", () => {
+	it("true is true", () => {
+		expect(true).toBe(true);
+	});
+});
